Fixed bug, that JPA NamedQuery expected experimnetKey as String, but was passed as a Long.
The ScheduledExperiment now return a long for the experiment key. At the
moment it still only refers to the hashCode() method, which returns an
Integer, but this might change later.

+++ org.sopeco.service/src/main/java/org/sopeco/service/persistence/entities/ScheduledExperiment.java
@@ -263,11 +263,11 @@ public class ScheduledExperiment implements Serializable {
 
 	/**
 	 * Calculates a unique the experiment key for this {@link ScheduledExperiment}.
-	 * At the moment only take the hash code of this object.
+	 * At the moment only take the hash code of this object.<br />
 	 * 
 	 * @return the experiment key
 	 */
-	public int getExperimentKey() {
+	public long getExperimentKey() {
 		return hashCode();
 	}
